1Mart-RAG-Bot/requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
google-generativeai
requests
python-dotenv
